refactor: restructe packages
