+++ docarray/__init__.py
@@ -1,15 +1,9 @@
 __version__ = '0.1.0'
 
 from docarray.array.array import DocumentArray
-from docarray.document.document import BaseDocument
-from docarray.predefined_document import Audio, Image, Mesh3D, PointCloud3D, Text
+from docarray.document_base.document import BaseDocument
 
 __all__ = [
     'BaseDocument',
     'DocumentArray',
-    'Image',
-    'Audio',
-    'Text',
-    'Mesh3D',
-    'PointCloud3D',
 ]
