--- docarray/document/__init__.py
@@ -1,5 +1,0 @@
-from docarray.document.any_document import AnyDocument
-from docarray.document.base_node import BaseNode
-from docarray.document.document import BaseDocument
-
-__all__ = ['AnyDocument', 'BaseDocument', 'BaseNode']
