--- docarray/document/mixins/__init__.py
@@ -1,3 +1,0 @@
-from docarray.document.mixins.proto import ProtoMixin
-
-__all__ = ['ProtoMixin']
