+++ docarray/document_base/__init__.py
@@ -1,0 +1,5 @@
+from docarray.document_base.any_document import AnyDocument
+from docarray.document_base.base_node import BaseNode
+from docarray.document_base.document import BaseDocument
+
+__all__ = ['AnyDocument', 'BaseDocument', 'BaseNode']
