+++ docarray/document_base/mixins/__init__.py
@@ -1,0 +1,3 @@
+from docarray.document_base.mixins.proto import ProtoMixin
+
+__all__ = ['ProtoMixin']
