+++ docarray/document_base/mixins/proto.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING, Any, Dict, Type, TypeVar
 
-from docarray.document.abstract_document import AbstractDocument
-from docarray.document.base_node import BaseNode
+from docarray.document_base.abstract_document import AbstractDocument
+from docarray.document_base.base_node import BaseNode
 
 if TYPE_CHECKING:
     from docarray.proto import DocumentProto, NodeProto
