+++ docarray/documents/__init__.py
@@ -1,0 +1,7 @@
+from docarray.documents.audio import Audio
+from docarray.documents.image import Image
+from docarray.documents.mesh import Mesh3D
+from docarray.documents.point_cloud import PointCloud3D
+from docarray.documents.text import Text
+
+__all__ = ['Text', 'Image', 'Audio', 'Mesh3D', 'PointCloud3D']
