--- docarray/predefined_document/__init__.py
@@ -1,7 +1,0 @@
-from docarray.predefined_document.audio import Audio
-from docarray.predefined_document.image import Image
-from docarray.predefined_document.mesh import Mesh3D
-from docarray.predefined_document.point_cloud import PointCloud3D
-from docarray.predefined_document.text import Text
-
-__all__ = ['Text', 'Image', 'Audio', 'Mesh3D', 'PointCloud3D']
