+++ docarray/typing/abstract_type.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, Type, TypeVar
 from pydantic import BaseConfig
 from pydantic.fields import ModelField
 
-from docarray.document.base_node import BaseNode
+from docarray.document_base.base_node import BaseNode
 
 if TYPE_CHECKING:
     from docarray.proto import NodeProto
