+++ docarray/typing/tensor/torch_tensor.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
     from docarray.proto import NdArrayProto, NodeProto
     from docarray.computation.torch_backend import TorchCompBackend
 
-from docarray.document.base_node import BaseNode
+from docarray.document_base.base_node import BaseNode
 
 T = TypeVar('T', bound='TorchTensor')
 ShapeT = TypeVar('ShapeT')
