+++ tests/units/array/test_generic_array.py
@@ -1,5 +1,5 @@
 from docarray import BaseDocument, DocumentArray
-from docarray.document import AnyDocument
+from docarray.document_base import AnyDocument
 
 
 def test_generic_init():
