+++ tests/units/document/proto/test_document_proto.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch
 
 from docarray import DocumentArray
-from docarray.document import BaseDocument
+from docarray.document_base import BaseDocument
 from docarray.typing import NdArray, TorchTensor
