+++ tests/units/typing/url/test_any_url.py
@@ -1,6 +1,6 @@
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.document.io.json import orjson_dumps
+from docarray.document_base.io.json import orjson_dumps
 from docarray.typing import AnyUrl
