+++ tests/units/typing/url/test_point_cloud_url.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.document.io.json import orjson_dumps
+from docarray.document_base.io.json import orjson_dumps
 from docarray.typing import PointCloud3DUrl
 from tests import TOYDATA_DIR
